ui: Refactor nodeGraphs page into multiple files
Splits the very large "nodeGraphs" component, which contained all graphs
on the cluster page, into multiple files - one for each "dashboard"
selectable on the cluster page. This should make it easier to find or
modify the graphs on individual dashboards; it should also serve as an
example of how to break up a complicated component into multiple files,
while still maintaining an approachable "public" interface in index.tsx.

+ Dashboards are defined in files under
/views/cluster/containers/nodeGraphs/dashboards. These are imported and
used in nodeGraph's index.tsx file.
+ Brought the formatting for the JSX of our graphs up to date with our
upcoming style guide (not currently enforced by linters)
+ Dashboard labels are defined only in one place now; previously, there
were several collections where string values needed to be kept in sync.
+ Removed the "GraphGroup" component which is no longer needed.
+ Removed the "Nodes" component, which was an unneeded wrapper that
provided the dropdown options to the node graphs page. These options are
now displayed directly by the graphs page.
